chor : moved to implement feature based structure
