+++ src/main/java/com/example/demo/features/mailSystem/dto/converter/SendererDtoConverter.java
@@ -1,8 +1,7 @@
-package com.example.demo.features.user.dto.converter;
+package com.example.demo.features.mailSystem.dto.converter;
 
-import com.example.demo.features.mailSystem.entity.Mail;
-import com.example.demo.features.user.entity.Senderer;
-import com.example.demo.features.user.dto.SendererDTO;
+import com.example.demo.features.mailSystem.entity.Senderer;
+import com.example.demo.features.mailSystem.dto.SendererDTO;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -10,13 +9,11 @@ import java.util.List;
 @Component
 public class SendererDtoConverter {
     public SendererDTO convert (Senderer from){
-        List<Long>mailIds = from.getMails().stream().map(Mail::getId).toList();
         return new SendererDTO(
                 from.getId(),
                 from.getFname(),
                 from.getLname(),
-                from.getEmail(),
-                mailIds
+                from.getEmail()
         );
     }
     public List<SendererDTO> convert (List<Senderer> from){
