+++ src/main/java/com/example/demo/features/mailSystem/dto/request/BaseSendererRequest.java
@@ -1,4 +1,4 @@
-package com.example.demo.features.user.dto.request;
+package com.example.demo.features.mailSystem.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
